Move load functions to a different file in the pki package
